(9/11/06) - title spectrometer snapshots with model name and snapshot# (9/11/06) - title dynamic spectrometer with "Photons emitted" (9/11/06) - add close icon to spectrometer snapshot (9/11/06) - remove start/stop, reset, and camera buttons from spectrometer snapshot (9/11/06) - new image for spectrometer, with title area (9/11/06) - image for spectrometer close button

+++ trunk/hydrogen-atom/src/edu/colorado/phet/hydrogenatom/control/GunOnOffControl.java
@@ -39,7 +39,7 @@ public class GunOnOffControl extends PhetPNode {
     // Class data
     //----------------------------------------------------------------------------
     
-    private static final Point2D BUTTON_OFFSET = new Point2D.Double( 33, 68 );
+    private static final Point2D BUTTON_OFFSET = new Point2D.Double( 33, 72 );
     
     //----------------------------------------------------------------------------
     // Instance data
